+++ src/test/kotlin/io/kjson/JSONSerializerTest.kt
@@ -2,7 +2,7 @@
  * @(#) JSONSerializerTest.kt
  *
  * kjson  Reflection-based JSON serialization and deserialization for Kotlin
- * Copyright (c) 2019, 2020, 2021 Peter Wall
+ * Copyright (c) 2019, 2020, 2021, 2022 Peter Wall
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -425,7 +425,7 @@ class JSONSerializerTest {
 
     @Test fun `should return Duration as JSONString`() {
         val duration = 2.hours
-        expect(JSONString("2h")) { JSONSerializer.serialize(duration) }
+        expect(JSONString("PT2H")) { JSONSerializer.serialize(duration) }
     }
 
     @Test fun `should return UUID as JSONString`() {